requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
PyYAML==6.0.2
user-agents==2.2.0
